+++ docs/astro.config.mjs
@@ -1,7 +1,6 @@
 import { $ } from 'zx'
 import { defineConfig } from 'astro/config'
 import starlight from '@astrojs/starlight'
-import preact from '@astrojs/preact'
 if (!process.env.VERCEL) await $`tsx sync-readme-to-pages.ts`
 
 // https://astro.build/config
@@ -75,7 +74,6 @@ export default defineConfig({
         },
       ],
     }),
-    preact(),
   ],
   // Process images with sharp: https://docs.astro.build/en/guides/assets/#using-sharp
   image: {